requirements.txt
kombu==5.3.4
asyncio-redis==0.16.0
amqp==5.2.0
backoff==2.2.1
celery==5.3.6
flower==2.0.1
openai==1.6.0
redis==5.0.1
requests==2.31.0
httpx==0.24.1
supabase==2.3.0
pydantic==2.5.2
python-dotenv==1.0.0
